refactor: split GameMapAreaStateLogic modules and add flame physical ratchet (#4183)
Extract shell, work-target, draft-projection, and province-action helpers
into dedicated libraries; keep GameMapAreaStateLogic as the stable forwarder
barrel. Add repo.app_flame_lib_physical_file_size (400-line ceiling with
shrink-only grandfather for region-map modules pending Slice B).

Refs #4183
