+++ SPEC/program/repo-lint.md
@@ -152,6 +152,7 @@
 | `tool/check_orders_test_support_loc.dart` | Physical-line ceilings for `packages/colonizethis_orders/test/orders/support/**/*.dart` (wave-5 support AC ≤13,950) and package `test/**/*.dart` (wave-5 package AC ≤15,800; Refs #4109); rule `repo.orders_test_support_loc` |
 | `tool/check_orders_test_shorthand_file_size.dart` | Physical-line ceiling (**180**) for every `*expectation_shorthand.dart` under `packages/colonizethis_orders/test/orders/support/**` (Refs #4109 wave 5); rule `repo.orders_test_shorthand_file_size` |
 | `tool/check_orders_lib_source_file_size.dart` | Ratchet ceiling for `packages/colonizethis_orders/lib/**` physical lines per file (wave-5 ≤400; shrink-only grandfather allowlist for `order_engine_validation.dart` and `work_order_validator.dart` still tracked by `repo.orders_file_size`; generated suffixes excluded); rule `repo.orders_lib_source_file_size` (Refs #4109) |
+| `tool/check_app_flame_lib_physical_file_size.dart` | Ratchet ceiling for `app/lib/features/game/flame/**` physical lines per file (wave-10 ≤400; shrink-only grandfather allowlist for region-map render modules and near-cap province-action states during transition; generated suffixes excluded); rule `repo.app_flame_lib_physical_file_size` (Refs #4183) |
 | `tool/check_orders_test_support_raw_game.dart` | Ceiling on `return Game(` / `=> Game(` under `test/orders/support/**` excluding `common/` (wave-4 migration ratchet; Refs #3971); rule `repo.orders_test_support_raw_game` |
 | `tool/check_orders_scenario_table_runner.dart` | Prefer scenario-table runners for new long imperative `test()` bodies under `packages/colonizethis_orders/test/**`; after wave-3 slice 13, baseline allow-all is off and remaining imperative suites are listed in `ordersPreferScenarioTablesAllowlist` / `REFACTOR_TRACE.md` (Refs #3949); rule `repo.orders_test_prefer_scenario_tables` |
 | `tool/check_combat_test_no_duplicate_fixtures.dart` | Forbid re-introducing a local copy of the shared game/world test fixtures anywhere under `packages/colonizethis_combat/test/**`: any file named `test_fixtures.dart` (the deleted byte-identical-modulo-doc-comment subset of `package:colonizethis_test/game_test_fixtures.dart`) or any test file that redeclares a top-level `TestFixtures` class (optionally `abstract`/`final`/`base`/`sealed`/`interface`/`mixin`) is a violation. Combat tests must import the canonical `package:colonizethis_test/game_test_fixtures.dart` factories (single source of truth) rather than duplicating them; references to the shared `TestFixtures` symbol are unaffected, and classes whose names merely contain `TestFixtures` (e.g. `MyTestFixturesBuilder`) are not flagged. Complements the testing-rule shared-fixtures policy (Refs #3716); rule `repo.combat_test_no_duplicate_fixtures` |
